% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biome_defaults}
\alias{biome_defaults}
\title{Biome Defaults.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
biome_defaults
}
\description{
Contains the default values for biomes, with rows containing variable names
and values, and columns containing biome codes. Selecting a specific column
provides the default calculator values for ghgvcR.
}
\keyword{datasets}

