% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{biome_write}
\alias{biome_write}
\title{Write Biome Data}
\usage{
biome_write(df, outdir, filename = "biome", format = c("json", "csv"))
}
\arguments{
\item{df}{the biome data.}

\item{outdir}{the directory to write the data to.}

\item{filename}{(character) name of file to write.}

\item{format}{(character) file format to write.}
}
\value{
TRUE if written with no errors.
}
\description{
importFrom jsonlite fromJSON
}

