% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghgvc.R
\name{calc_ghgv}
\alias{calc_ghgv}
\title{GHGVC}
\usage{
calc_ghgv(eco, is_xml = FALSE, output_filename = "ghgv.csv",
  save_output = FALSE, plot_filename = "ghgv_plot.png",
  plot_units = c("co2", "mi"), save_plots = FALSE)
}
\arguments{
\item{is_xml}{(logical) indicating that eco_json is xml instead of json.}

\item{output_filename}{(character) name of file to write.}

\item{save_output}{(logical) indicating whether or not to write the results.}

\item{plot_filename}{(character) name of plot to write.}

\item{plot_units}{(character) units to plot in. May be contain multiple values.}

\item{save_plots}{(logical) indicating whether or not to save the plot images.
the results.}

\item{eco_json}{(json string) A JSON list of configuration settings and data parameters.}
}
\value{
List of GHGVC results for each location specified in \code{eco_json}.
}
\description{
Greenhouse Gas Value Calculator
}
\details{
R function implementation of GHGV_calculator_web.m
}
\author{
Chris Schauer, David LeBauer, Nicholas Potter
}
