% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{decay}
\alias{decay}
\title{Decay function for rates over time.}
\usage{
decay(r, t)
}
\arguments{
\item{r}{a vector of decay rates.}

\item{t}{a sequential vector of time steps.}
}
\value{
a matrix of decay fluxes for each time moment.
}
\description{
Decay function for rates over time.
}

