% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_ghg_params}
\alias{extract_ghg_params}
\title{Extract GHG parameters into a matrix.}
\usage{
extract_ghg_params(ecosystem, include_anth = TRUE)
}
\arguments{
\item{ecosystem}{a named list of ecosystem traits.}

\item{include_anth}{boolean whether to include Anthro flux}
}
\value{
a matrix of greenhouse gas parameters.
}
\description{
Extract GHG parameters into a matrix.
}
