% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_pool_params}
\alias{extract_pool_params}
\title{Extract pool parameters into a matrix.}
\usage{
extract_pool_params(ecosystem)
}
\arguments{
\item{ecosystem}{a list of ecosystem traits.}
}
\value{
a matrix.
}
\description{
Extract pool parameters into a matrix.
}
