% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fao_biomes}
\alias{fao_biomes}
\title{FAO Biomes.}
\format{A data frame with 15 rows and 3 variables:
\describe{
  \item{FAO}{FAO vegetation code.}
  \item{GX}{Grassland biome code.}
  \item{APX}{Pasture biome code.}
}}
\source{
\url{?}
}
\usage{
fao_biomes
}
\description{
Provides the grassland and pasture biome codes for a given FAO code.
}
\keyword{datasets}

