% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{get_biome}
\alias{get_biome}
\title{Get biome data from a netCDF file.}
\usage{
get_biome(ncfile, latitude, longitude)
}
\arguments{
\item{ncfile}{name of the netCDF file.}

\item{latitude}{the latitude of the biome.}

\item{longitude}{the longitude of the biome.}
}
\value{
JSON of biome data.
}
\description{
Get biome data from a netCDF file.
}

