% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{get_ncdf}
\alias{get_ncdf}
\title{Get biome data from a netCDF file.}
\usage{
get_ncdf(data_dir, ncfile, latitude, longitude, variables = "all")
}
\arguments{
\item{data_dir}{directory containing the data file.}

\item{ncfile}{name of the netCDF file.}

\item{latitude}{the latitude.}

\item{longitude}{the longitude.}

\item{variables}{the names of variables to select. If missing then all are 
returned.}
}
\value{
JSON of netcdf results.
}
\description{
Get biome data from a netCDF file.
}
