% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{get_ncdf_vars}
\alias{get_ncdf_vars}
\title{Get netcdf variable names.}
\usage{
get_ncdf_vars(file_name)
}
\arguments{
\item{file_name}{file name (relative to working directory).}
}
\value{
list of variable names.
}
\description{
Wrapper to simplify checking variables in a netcdf file.
}
