\name{ghgvc}
\alias{ghgvc}
\title{GHGVC}
\usage{
ghgvc(options, ecosystem_data)
}
\arguments{
  \item{options}{}

  \item{ecosystem_data}{}
}
\value{
GHGVC result
}
\description{
Greenhouse Gas Value Calculator
}
\details{
R function implementation of GHGV_calculator_web.m
}
\author{
Chris Schauer, David LeBauer
}

