% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ghgvc_plot}
\alias{ghgvc_plot}
\title{Plot Greenhouse Gas values.}
\usage{
ghgvc_plot(df, outdir, years, save = TRUE, savefile = "output.svg")
}
\arguments{
\item{df}{a data.frame of output from \code{ghgvc()}.}

\item{outdir}{the directory to save the plot image.}

\item{years}{number of years considered in the analysis.}

\item{save}{boolean to save plot as an image.}

\item{savefile}{name of svg file to save.}
}
\value{
a ggplot2 plot object.
}
\description{
Plot Greenhouse Gas values.
}

