% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ghgvc_subplot}
\alias{ghgvc_subplot}
\title{Plot ghgcv subplot.}
\usage{
ghgvc_subplot(vars, data, baseplot)
}
\arguments{
\item{vars}{a list of variables to plot.}

\item{baseplot}{a ggplot2 object as the base plot.}

\item{longdata}{long form data.frame of data from \code{ghgcv()}.}
}
\value{
a ggplot2 object.
}
\description{
Creates a ggplot2 object using long-form data from \code{ghgcv_plot()} and
a set of variables to plot over.
}

