% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ghgvc_write}
\alias{ghgvc_write}
\title{Write Greenhouse Gas value calculator output to file.}
\usage{
ghgvc_write(df, outdir, filename = "output", format = c("json", "csv"))
}
\arguments{
\item{df}{data.frame of output from \code{ghgvc()}.}

\item{outdir}{the directory to write the data to.}

\item{filename}{(character) name of file to write.}

\item{format}{(character) file format to write.}
}
\value{
TRUE if written with no errors.
}
\description{
Write Greenhouse Gas value calculator output to file.
}

