% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{json_to_df}
\alias{json_to_df}
\title{Convert GHGVC json results to data.frame.}
\usage{
json_to_df(json)
}
\arguments{
\item{res}{a json object containing results from \code{ghgvc()}.}
}
\value{
a data frame of ghgvc results.
}
\description{
Convert GHGVC json results to data.frame.
}

