% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{kinetic_decay}
\alias{kinetic_decay}
\title{Decay Kinetics.}
\usage{
kinetic_decay(time_vector)
}
\arguments{
\item{time_vector}{a vector of time moments.}
}
\value{
a matrix of decay kinetics for each time moment.
}
\description{
Decay Kinetics.
}
