% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{koppen_biomes}
\alias{koppen_biomes}
\title{Koppen-Geiger Climate Zones and Biome Codes.}
\format{A data frame with 31 rows and 13 columns.
\describe{
  \item{Zone}{Climate zone.}
  \item{...}{Vegetation types.}
}}
\usage{
koppen_biomes
}
\description{
Data set of Koppen-Geiger climate zones and the biome codes that are 
associated with different vegetation types. For looking up the biome code
when provided with a climate zone and a vegetation type.
}
\keyword{datasets}

