% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_vegtypes}
\alias{map_vegtypes}
\title{Vegetation types by map value.}
\format{a data frame with 118 rows and 15 columns
\describe{
  \item{Map}{Map from which the value was drawn.}
  \item{value}{Map value.}
  \item{Category}{Vegetation type category}
  \item{...}{Vegetation types (int).}
}}
\usage{
map_vegtypes
}
\description{
Boolean indicators of vegetation types available an each map value.
}
\keyword{datasets}

