% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ghgv}
\alias{plot_ghgv}
\title{Plot Greenhouse Gas values.}
\usage{
plot_ghgv(df, years = 50, units = c("co2", "mi"), crv_to_miles = 1.86)
}
\arguments{
\item{df}{a data.frame of output from \code{ghgvc()}.}

\item{years}{(integer) years considered in the analysis.}

\item{units}{(character) either co2 or mi to plot in co2 equivalents or miles.}

\item{crv_to_miles}{(numeric) conversion factor for crv to miles, default 1.86.}
}
\value{
a ggplot2 plot object.
}
\description{
Plot Greenhouse Gas values.
}
