% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remap_range}
\alias{remap_range}
\title{Remap latitude and longitude ranges.}
\usage{
remap_range(input, input_min, input_max, output_min, output_max)
}
\arguments{
\item{input}{the value to remap.}

\item{input_min}{minimum input value.}

\item{input_max}{maximum input value.}

\item{output_min}{minimum output value.}

\item{output_max}{maximum output value.}
}
\description{
Remap latitude and longitude ranges.
}
