% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{str2LogicalOrNumeric}
\alias{str2LogicalOrNumeric}
\title{Convert string to logical or numeric.}
\usage{
str2LogicalOrNumeric(string)
}
\arguments{
\item{x}{(character) a string}
}
\value{
a logical or numeric value
}
\description{
Convert string to logical or numeric.
}
\examples{
\dontrun{
  str2LogicalOrNumeric("TRUE") # -> TRUE
  str2LogicalOrNumeric("5") # -> 5
  str2LogicalOrNumeric("NaN") # -> "NaN"
}
}
