% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vegtype_order}
\alias{vegtype_order}
\title{Determine factor order of a vegtype.}
\usage{
vegtype_order(vegtypes)
}
\arguments{
\item{vegtypes}{(character list) list or array of vegtype names.}
}
\value{
(integer list) order factor.
}
\description{
Provides an integer used to order biomes by vegtype.
}
