% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_json}
\alias{write_json}
\title{Writes a json string to a json or csv file.}
\usage{
write_json(json, outdir, filename = "json", format = c("json", "csv"))
}
\arguments{
\item{outdir}{the directory to write the data to.}

\item{filename}{(character) name of file to write.}

\item{format}{(character) file format to write.}

\item{df}{the data.}
}
\value{
TRUE if written with no errors.
}
\description{
importFrom jsonlite toJSON fromJSON validate
}

