% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_output}
\alias{write_output}
\title{Writes a json string to a json or csv file.}
\usage{
write_output(json, output_filename)
}
\arguments{
\item{json}{a json object of the ghgv output.}

\item{output_filename}{(character) the filename to output, extension included.}
}
\value{
TRUE if written with no errors.
}
\description{
Writes a json string to a json or csv file.
}

