% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_plot}
\alias{write_plot}
\title{Writes a json string to a json or csv file.}
\usage{
write_plot(plt, plot_filename)
}
\arguments{
\item{plt}{a ggplot2 object.}

\item{plot_filename}{(character) the name of the plot file to save, extension included.}
}
\value{
TRUE if written with no errors.
}
\description{
Writes a json string to a json or csv file.
}

