% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghgvc.R
\name{calc_ghgv}
\alias{calc_ghgv}
\title{GHGVC}
\usage{
calc_ghgv(eco_json, output_filename = "ghgv.csv", save_output = FALSE,
  plot_filename = "ghgv_plot.png", plot_units = c("co2", "mi"),
  save_plots = FALSE)
}
\arguments{
\item{eco_json}{(json string) A JSON list of configuration settings and data parameters.}

\item{output_filename}{(character) name of file to write.}

\item{plot_filename}{(character) name of file to write.}

\item{output_format}{(character) format of output file, either json or csv.}

\item{write_output}{(logical) indicating whether or not to write the results.}

\item{plot_format}{(character) format of output file, either json or csv.}

\item{write_plot}{(logical) indicating whether or not to write the results.
the results.}
}
\value{
List of GHGVC results for each location specified in \code{eco_json}.
}
\description{
Greenhouse Gas Value Calculator
}
\details{
R function implementation of GHGV_calculator_web.m
}
\author{
Chris Schauer, David LeBauer, Nicholas Potter
}

