% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{get_biome}
\alias{get_biome}
\title{Get biome data from netcdf and json files and return and/or write a config 
file.}
\usage{
get_biome(latitude, longitude, data_dir = "/home/ghgvcr/data/",
  output_filename = "biome", output_format = c("json", "csv"),
  write_data = TRUE)
}
\arguments{
\item{latitude}{the selected latitude.}

\item{longitude}{the selected longitude.}

\item{data_dir}{path of the netcdf and map csv data files.}

\item{output_filename}{name of file to write (without extension). Only needed if write is TRUE.}

\item{output_format}{format to save data in.}

\item{write_data}{boolean whether to write the data.}
}
\value{
JSON of biome data.
}
\description{
Get biome data from netcdf and json files and return and/or write a config 
file.
}

