% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome.R
\name{get_biome}
\alias{get_biome}
\title{Get biome data from netcdf and json files and return and/or write a config 
file.}
\usage{
get_biome(latitude, longitude, biome_defaults_file, netcdf_dir, mapdata_dir,
  output_dir, output_filename = "biome", output_format = c("json", "cvs"),
  write_data = TRUE)
}
\arguments{
\item{latitude}{the selected latitude.}

\item{longitude}{the selected longitude.}

\item{biome_defaults_file}{full path of the name-indexed ecosystem json file.}

\item{netcdf_dir}{full path to the directory containing the netcdf data files.}

\item{output_dir}{full path of the directory to write results to.}

\item{output_filename}{name of file to write (without extension).}

\item{output_format}{format to save data in.}

\item{write}{boolean whether to write the data.}
}
\value{
JSON of biome data.
}
\description{
Get biome data from netcdf and json files and return and/or write a config 
file.
}

