% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghgvc.R
\name{ghgvc}
\alias{ghgvc}
\title{GHGVC}
\usage{
ghgvc(config)
}
\arguments{
\item{config}{A list of configuration settings and data parameters.}
}
\value{
List of GHGVC results for each location specified in \code{config}.
}
\description{
Greenhouse Gas Value Calculator
}
\details{
R function implementation of GHGV_calculator_web.m
}
\author{
Chris Schauer, David LeBauer, Nicholas Potter
}

