% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ghgv}
\alias{plot_ghgv}
\title{Plot Greenhouse Gas values.}
\usage{
plot_ghgv(df, output_dir, years = 50, save = TRUE,
  savefile = "output.svg")
}
\arguments{
\item{df}{a data.frame of output from \code{ghgvc()}.}

\item{output_dir}{the directory to save the plot image.}

\item{years}{number of years considered in the analysis.}

\item{save}{boolean to save plot as an image.}

\item{savefile}{name of svg file to save.}
}
\value{
a ggplot2 plot object.
}
\description{
Plot Greenhouse Gas values.
}

